finished

+++ src/controllers/EventoController.js
@@ -28,7 +28,6 @@ router.get("/" , async (req, res) => {
   try {
     if (esFecha(Evento.startDate) || Evento.startDate == undefined) {
       const allEvents = await EventService.getEventByFilter(Evento, limit, offset);
-      console.log(allEvents);
       return res.send(allEvents);
     } else {
       return res.json("error en los filtros ingresados");
@@ -45,13 +44,14 @@ router.delete("/:id", AuthMiddleware , async (req, res) => {
   try {
     const ev=await EventService.getEventById(id)
     const tags=await EventService.getTags(id)
+    const enrollment=await EventService.getEventEnrollment(id)
     if(ev!=null){
-    if (tags!=null) {
+    if (tags==null && enrollment==null) {
       
       const eliminado=await EventService.DeleteEvent(id);
       return res.status(200).json(eliminado)
     }else{
-    return res.status(400).send("No se puede eliminar porque tiene tags");4
+    return res.status(400).send("No se puede eliminar porque tiene tags o gente inscripta");
     }
   }else{
     return res.status(404).send("No existe la id")
@@ -78,7 +78,6 @@ router.post("/",AuthMiddleware, async (req, res) => {
   try {
     const evLoc  =await evLocService.getEventLocationById(Evento.id_event_location)
     if (Evento.name!=null && Evento.description!=null && Evento.id_event_category!=null && Evento.id_event_location!=null && Evento.start_date!=null && Evento.duration_in_minutes!=null && Evento.price!=null && Evento.enabled_for_enrollment!=null && Evento.max_assistance!=null) {
-      console.log(evLoc);
       if (evLoc.max_capacity>Evento.max_assistance) {
         if (Evento.price>0 && Evento.duration_in_minutes>0 ) {
           const respuesta = await EventService.InsertEvento(Evento);;
@@ -114,10 +113,8 @@ router.put("/",AuthMiddleware, async (req, res) => {
   Evento.id_event_location=req.body.id_event_location;
     
   try {
-    console.log(Evento.id);
     const ev=await EventService.getEventById(Evento.id)
     const evLoc  =await evLocService.getEventLocationById(Evento.id_event_location)
-    console.log(ev);
     if (ev!=null) {
       
     if (Evento.name!=null && Evento.description!=null && Evento.id_event_category!=null && Evento.id_event_location!=null && Evento.start_date!=null && Evento.duration_in_minutes!=null && Evento.price!=null && Evento.enabled_for_enrollment!=null && Evento.max_assistance!=null) {
@@ -204,7 +201,7 @@ router.post("/:id/enrollment", AuthMiddleware , async (req, res) => {
   }
 });
 
-router.put("/:id/enrollment",AuthMiddleware,async (req, res) => {
+router.patch("/:id/enrollment",AuthMiddleware,async (req, res) => {
   const idEvento = req.params.id;
   const rating = req.query.rating;
   try {
